Some typos in README.md
The link to Luise Freese Twitter account is incorrect,  https://tiwtter.com/LuiseFreese > https://twitter.com/luisefreese.
The Markdown definition for the Checkbox link has double parents.
The link to the GenericInputFrame was incorrect. The  Markdown filename is wrong GerenericInputFrame.md > GenericInputFrame.md.
The link to the TextInput was incorrect. The Markdown filename is wrong Textinput.md > TextInput.md.